Move ISingleChargeMs1Feature to MassSpectrometry
The interface describes an MS1 deconvolution result and is the input
contract for the upcoming Deconvoluter.PairPrecursorsToMs2 join. That
method lives in MassSpectrometry.Deconvolution; Readers references
MassSpectrometry rather than the reverse, so the interface had to live
in MassSpectrometry for the dependency direction to work.

Relocate the file via `git mv` (history preserved), re-namespace to
MassSpectrometry, and add `using MassSpectrometry;` to the five Readers
files that implement or reference it. No behavior change. No external
consumer references the interface today (grep-checked on MetaMorpheus,
ProteaseGuru, ProteoformExplorer), so the move is invisible outside
this PR.

Doc tightened to reflect that pairing requires both an RT-window check
and an isolation-window check, not just an RT-apex check.
