+++ mzLib/Readers/ExternalResults/IndividualResultRecords/DinosaurFeature.cs
@@ -2,6 +2,7 @@
 using System.Text;
 using CsvHelper.Configuration;
 using CsvHelper.Configuration.Attributes;
+using MassSpectrometry;
 
 namespace Readers
 {
