+++ mzLib/Readers/ExternalResults/IndividualResultRecords/Ms1Feature.cs
@@ -3,6 +3,7 @@ using System.Text;
 using Chemistry;
 using CsvHelper.Configuration;
 using CsvHelper.Configuration.Attributes;
+using MassSpectrometry;
 
 namespace Readers
 {
