+++ mzLib/Readers/ExternalResults/ResultFiles/DinosaurTsvFile.cs
@@ -1,4 +1,5 @@
 ﻿using CsvHelper;
+using MassSpectrometry;
 
 namespace Readers
 {
