+++ mzLib/Readers/ExternalResults/ResultFiles/Ms1FeatureFile.cs
@@ -1,4 +1,5 @@
 ﻿using CsvHelper;
+using MassSpectrometry;
 
 namespace Readers
 {
